packages/corp-os-meta/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "corp_os_meta"
version = "1.0.0"
description = "Shared metadata schema and taxonomy for corp-by-os agent ecosystem"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "json-repair>=0.30",
    "python-dotenv>=1.0",
]

[tool.setuptools.package-data]
corp_os_meta = ["data/*.yaml"]

[project.scripts]
corp-meta = "corp_os_meta.cli:main"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
